ai-web/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
faiss-cpu
numpy
SQLAlchemy
psycopg2-binary
alembic
